pyproject.toml
[tool.poetry]
name = "finance"
version = "0.0.1"
description = ""
authors = ["Sergey Parhomenko <sergey@parhomenko.net>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.10.1" # renovate github-tags:python/cpython:^v(?<version>.*)$
"pdfminer.six" = "20211012"
beautifulsoup4 = "4.10.0"
bpylist2 = "4.0.1"
chromedriver-binary-auto = "0.1.1"
cryptography = "36.0.1"
dacite = "1.6.0"
inflection = "0.5.1"
more-itertools = "8.12.0"
pyotp = "2.6.0"
requests = "2.26.0"
selenium = "4.1.0"
srp = "1.0.18"

[tool.poetry.dev-dependencies]
autopep8 = "1.6.0"
black = {version = "21.12b0", allow-prereleases = true}
coverage = "6.2"
flake8 = "4.0.1"
flake8-black = "0.2.3"
flake8-isort = "4.1.1"
mypy = "0.930"
pytest = "6.2.5"
types-beautifulsoup4 = "4.10.7"
types-requests = "2.26.3"
types-selenium = "3.141.6"
wemake-python-styleguide = "0.16.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 200
target-version = ["py310"]

[tool.isort]
line_length = 200
profile = "black"
py_version = 310

[tool.mypy]
mypy_path = "$MYPY_CONFIG_FILE_DIR/stubs"
strict = true
disallow_any_decorated = true
disallow_any_explicit = true
disallow_any_expr = true
disallow_any_unimported = true
warn_unreachable = true
